refactor(ranking): share RankingSortButtons, drop CoverageSortControls
Extract the two sort toggles (alphabetical/value + ascending/descending) into a
new RankingSortButtons widget in ranking_filter_controls.dart, used by
RankingFilterControls, RailwayCoverageView and RailAreaRankingPage. Remove the
now-redundant CoverageSortControls widget/file. Behaviour unchanged (enabled
defaults to true; the coverage views still pass sortingEnabled).
